Refactor : Cloud 클래스 리팩터링
