+++ Assets/Prefab/Platform Star.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3447c827c954d374c95c16051c4e175f
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
