+++ Assets/Scripts/Object/Cloud/DisappearedCloud.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0abd27a553a849c46a2949f82b33fe8c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
